Create Edge Detection : Separating overlapping objects
Create a script that:
Converts the image to grayscale
Segments the grayscale image and saves the result as BW. Your mask should have the shapes in the foreground with a dark background.
Detects the edges of the grayscale image using the sobel method and saves the result as imgEdges. Specify your input options to skip the edge-thinning step to make your later segmentation more effective.
Uses logical indexing to remove the edges from the original segmentation and saves the result as newBW. 
The original image, "DetectingEdgesSlide.png" is available in the course files. You are encouraged to work in MATLAB to inspect the images and choose thresholds that are appropriate for your segmentations. Your final segmentation should have 6 distinct regions instead of 5.
Uncomment the code available at the bottom of the script to display your results. Your edges will be enhanced to make them easier to see on your screen. You will learn how to do this later in this course.